+++ Changelog.md
@@ -2,6 +2,7 @@
 
    * Fixed player escorts trying to respond to distress signals against the
      player
+   * Fixed several missions using inexistent audio.soundPlay function
 
 
 ## v0.13.0